Fix compilation without CAF

+++ drivers/net/wireguard/ratelimiter.c
@@ -183,9 +183,9 @@ int wg_ratelimiter_init(void)
 	 * dependent on RAM. This calculation here comes from there.
 	 */
 	table_size = (totalram_pages() > (1U << 30) / PAGE_SIZE) ? 8192 :
-		max_t(unsigned long, 16, roundup_pow_of_two(
-			(totalram_pages() << PAGE_SHIFT) /
-			(1U << 14) / sizeof(struct hlist_head)));
+             max_t(unsigned long, 16, roundup_pow_of_two(
+                 ((unsigned long)totalram_pages() << PAGE_SHIFT) /
+                 ((1U << 14) / sizeof(struct hlist_head))));
 	max_entries = table_size * 8;
 
 	table_v4 = kvcalloc(table_size, sizeof(*table_v4), GFP_KERNEL);
